Don't link non-existing shared libraries on static-only platforms, bug #394877
(Portage version: 2.2.01.20043-prefix/svn/Darwin i386, signed Manifest commit with key C5C74E89)

+++ sys-libs/ncurses/ncurses-5.9-r1.ebuild
@@ -212,7 +212,9 @@ src_install() {
 	if use unicode ; then
 		gen_usr_ldscript libncursesw$(get_libname)
 	fi
-	ln -sf libncurses$(get_libname) "${ED}"/usr/$(get_libdir)/libcurses$(get_libname) || die
+	if ! tc-is-static-only ; then
+		ln -sf libncurses$(get_libname) "${ED}"/usr/$(get_libdir)/libcurses$(get_libname) || die
+	fi
 	use static-libs || rm "${ED}"/usr/$(get_libdir)/*.a
 
 #	if ! use berkdb ; then
